feat(webapp): gather errors in a module
This commit introduces the module `page/error.rs` that defines a few
components to display various error in a uniform way.
