+++ typhon-webapp/src/pages/mod.rs
@@ -1,4 +1,5 @@
 pub mod dashboard;
+pub mod error;
 pub mod evaluation;
 pub mod jobset;
 pub mod login;
@@ -6,6 +7,7 @@ pub mod project;
 pub mod projects;
 
 pub(crate) use dashboard::Dashboard;
+pub(crate) use error::*;
 pub(crate) use evaluation::Evaluation;
 pub(crate) use jobset::Jobset;
 pub(crate) use login::Login;
